contactControllers

+++ controllers/contactsControllers.js
@@ -3,12 +3,11 @@ import HttpError from "../helpers/HttpError.js";
 
 async function getAllContacts(req, res, next) {
   const { _id: owner } = req.user;
-  //const { page = 1, limit = 10, favorite } = req.query;
+  //const { page = 1, limit = 10} = req.query;
   //const skip = (page - 1) * limit;
   try {
-    //  const result = await Contact.find({ owner }, { skip, limit });
+    //const result = await Contact.find({ owner }, { skip, limit });
     const result = await Contact.find({ owner });
-    // const result = await Contact.find({ ownerId: req.user.id });
     res.status(200).send(result);
   } catch (error) {
     next(error);
@@ -19,7 +18,10 @@ async function getOneContact(req, res, next) {
   try {
     const result = await Contact.findById(id);
     if (result === null) {
-      throw HttpError(404);
+      throw HttpError(404, "Contact is not found");
+    }
+    if (result.owner.toString() !== req.user.id) {
+      throw HttpError(404, "Contact is not found");
     }
     res.status(200).send(result);
   } catch (error) {
@@ -30,7 +32,7 @@ async function getOneContact(req, res, next) {
 async function deleteContact(req, res, next) {
   const { id } = req.params;
   try {
-    const deletedContact = await Contact.findByIdAndDelete(id);
+    const deletedContact = await Contact.findByIdAndDelete(id, { owner });
     if (!deletedContact) {
       throw HttpError(404, "Contact is not found");
     }
@@ -40,16 +42,6 @@ async function deleteContact(req, res, next) {
   }
 }
 
-// async function createContact(req, res, next) {
-//   const { _id: owner } = req.user;
-//   try {
-//     const newContact = await Contact.create(...req.body, owner);
-//     console.log(newContact);
-//     res.status(201).json(newContact);
-//   } catch (error) {
-//     next(error);
-//   }
-// }
 async function createContact(req, res, next) {
   const { _id: owner } = req.user;
   try {
@@ -63,9 +55,14 @@ async function createContact(req, res, next) {
 async function updateContact(req, res, next) {
   const { id } = req.params;
   try {
-    const contactToUpdate = await Contact.findByIdAndUpdate(id, req.body, {
-      new: true,
-    });
+    const contactToUpdate = await Contact.findByIdAndUpdate(
+      id,
+      { owner },
+      req.body,
+      {
+        new: true,
+      }
+    );
     if (!contactToUpdate) {
       throw HttpError(404, "Contact is not found");
     }
@@ -78,7 +75,9 @@ async function updateContact(req, res, next) {
 async function updateFavorite(req, res, next) {
   const { id } = req.params;
   try {
-    const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
+    const result = await Contact.findByIdAndUpdate(id, { owner }, req.body, {
+      new: true,
+    });
     if (!result) {
       throw HttpError(404, "Contact is not found");
     }
